requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
matplotlib
pyyaml

# optional (only if you want real Marigold)
diffusers
transformers
accelerate
safetensors
